feat(day06): this took me a while because of a stupid mistake
I solved it but got stuck because I was trying to return early from my reduce under a certain condition, however I was forgetting where I was and was thinking I could "return" early but it wasn't. I had it the right way originally but I refactored it because I'm so used to breaking up if/else conditionals in other languages like that!